[kernel] ext4 INFO: recovery required on readonly filesystem
Input:
[   11.868857] EXT4-fs (sda1): mounted filesystem with ordered data mode. Opts: (null)
[   17.865885] EXT4-fs (sda1): re-mounted. Opts: (null)
[   11.837260] EXT4-fs (sda1): mounted filesystem with ordered data mode. Opts: (null)
[   18.062820] EXT4-fs (sda1): re-mounted. Opts: (null)
[   11.927606] EXT4-fs (sda1): INFO: recovery required on readonly filesystem
[   11.935178] EXT4-fs (sda1): write access will be enabled during recovery
[   12.456601] EXT4-fs (sda1): recovery complete
[   12.463576] EXT4-fs (sda1): mounted filesystem with ordered data mode. Opts: (null)
[   18.922399] EXT4-fs (sda1): re-mounted. Opts: (null)
[   12.066404] EXT4-fs (sda1): mounted filesystem with ordered data mode. Opts: (null)
[   19.829214] EXT4-fs (sda1): re-mounted. Opts: (null)

Old output:
 WARNING:  Kernel Errors Present
    INFO: recovery required on readonly filesyste ...:  1 Time(s)

New output:
 WARNING:  Ext4 Errors Present
    sda1:
       INFO: recovery required on readonly filesystem
       write access will be enabled during recovery
       recovery complete

The mounted / re-mounted lines are normal/noise.
The INFO line either means what it says, or it hints that recovery is going to happen now.

When it hints that the system is going to recover on its own, it's important to see that
it has recovered.

When it means what it says, the user should actually do something.
For this case, I think seeing all of the lines until the end of time is probably better than
just seeing "something" without a hint about the volume.

We can revisit that use case if someone has an example of it.

Note: I think that this code could be applied to EXT3, but I'm not sure enough people use
it for that to matter.