[exim] tls handling
Here's example output for a single day:

 --- TLS Connection Issues
   gnutls_handshake: A packet with illegal or unsupported version was received: 2 Time(s)
   gnutls_handshake: An unexpected TLS packet was received: 10 Time(s)
   gnutls_handshake: No supported cipher suites have been found: 2 Time(s)
   gnutls_handshake: The TLS connection was non-properly terminated: 1 Time(s)
   recv: An unexpected TLS packet was received: 1 Time(s)
   recv: Error in the pull function: 3 Time(s)
   recv: The TLS connection was non-properly terminated: 5 Time(s)
   send: Error in the push function: 82 Time(s)
   send: The specified session has been invalidated for some reason: 8 Time(s)

   From these hosts:
   (NAM01-BN3-obe.outbound.protection.outlook.com): 13 Time(s)
   (NAM01-BY2-obe.outbound.protection.outlook.com): 7 Time(s)
   (NAM01-SN1-obe.outbound.protection.outlook.com): 7 Time(s)
   (NAM02-BL2-obe.outbound.protection.outlook.com): 9 Time(s)
   (NAM02-CY1-obe.outbound.protection.outlook.com): 8 Time(s)
   (NAM02-SN1-obe.outbound.protection.outlook.com): 4 Time(s)
   (NAM03-BY2-obe.outbound.protection.outlook.com): 12 Time(s)
   (NAM03-CO1-obe.outbound.protection.outlook.com): 15 Time(s)
   (NAM03-DM3-obe.outbound.protection.outlook.com): 7 Time(s)
   116.252.0.155: 1 Time(s)
   139.60.160.154: 9 Time(s)
   27.16.169.132: 1 Time(s)
   flower.census.shodan.io: 10 Time(s)
   mail.a211.static.mgsend.net: 2 Time(s)
   researchscan319.eecs.umich.edu: 1 Time(s)
   scan-7.security.ipip.net: 6 Time(s)
   smtp-out-so.shaw.ca: 2 Time(s)

The code to switch from reverseHost to hostLabel reduces the number of lines for the outlook.com case.

Note: technically, neither "reverseHost" nor "hostLabel" are trustworthy -- they'd only be trustworthy
if we did a forward lookup on them to confirm that they resolve back to the same IP address.
But if we don't collapse things, the result is hundreds of lines of IP addresses which aren't remotely
helpful. We assume that in general hosts are vaguely trustworthy.

Note: shodan.io intentionally sends a bogus hostLabel in one case. As its bogus hostLabel does
not match the domain of the reverseHost, it's discarded. But, again, it could have its rDNS record
contain similar garbage.

I've left the ()s around the label cases to make it clearer that this is a different data field than
the normal hosts -- and because of sorting, they sort separately.